djangoserver/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-simpleui==4.0.3
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
